Выгрузка релиза

+++ kernel/include/user/env.h
@@ -15,12 +15,12 @@ typedef struct env
   int OS_VERSION_PATCH;                  ///< Патч
   size_t Link_Display;                ///< Ссылка на виртуальный экран
   size_t Link_Time;                   ///< Ссылка на время
-  int Display_W;                        ///< Длина экрана
-  int Display_H;                        ///< Высота экрана
+  size_t Display_W;                        ///< Длина экрана
+  size_t Display_H;                        ///< Высота экрана
   int Display_B;                        ///< Смещение экрана
-  int Display_P;                        ///< Глубина экрана
-  int Display_S;                        ///< Размер буфера
-  int Ticks;                            ///< Количество текущих тиков
+  size_t Display_P;                        ///< Глубина экрана
+  size_t Display_S;                        ///< Размер буфера
+  size_t Ticks;                            ///< Количество текущих тиков
   size_t RAM_Install;                 ///< Установлено ОЗУ
   size_t RAM_Used;                    ///< Использовано ОЗУ
   size_t RAM_Free;                    ///< Свободно ОЗУ
